Transferred Abstract Factory code from GoF
I kept naming conventions as is for now; also took a few liberties as to how some of these functions would be implemented. The code from the textbook isn't complete, but I don't think it is necessarily meant to be -- it is adequately complete such that it demonstrates the intended functionality for the desired concept.